requirements.txt
asgiref==3.4.1
Django==2.2.24
django-3-jet==1.0.8
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing==3.7.4.3
typing-extensions==3.10.0.2
